requirements.txt
streamlit==1.32.2
openai==1.76.0
python-dotenv==1.1.0
langchain==0.3.24
langchain-openai==0.3.14
langchain-community==0.3.23
langchain-core==0.3.56
langchain-experimental==0.3.4
langchain-teddynote==0.3.45
langchain-text-splitters==0.3.8
langgraph==0.4.3
langgraph-checkpoint==2.0.25
langgraph-prebuilt==0.1.8
langgraph-sdk==0.1.66
langsmith==0.3.38
faiss-cpu==1.11.0
numpy==1.26.4
pandas==2.2.3
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0

# Testing dependencies
pytest==8.0.2
pytest-cov==4.1.0
pytest-asyncio==0.23.5
pytest-mock==3.12.0
